requirements.txt
python_pptx==0.6.18
